p2p_chat/requirements.txt
Pillow
cryptography

